requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pytest==6.2.5
requests==2.26.0
kafka-python==2.0.2
python-dotenv==0.19.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.25
alembic==1.7.3
coverage==5.5
